requirements.txt
tensorflow
flask
flask-cors
numpy<2.0.0
opencv-python
scikit-learn
pillow
werkzeug
